[hyun.k] 방탈출 인증 관리 1 & 2단계 (#19)
* feat(auth): 토큰 발급 기능 구현

* feat(auth): 내 정보 조회 기능 구현

* feat(auth): Auth 테스트 구현

* feat(auth): interceptor 구현

* feat(auth): resolver 구현

* feat(auth): 예약 삭제 구현

* feat(auth): reservation 테스트 구현

* feat(auth): exceptionHandler 제거

* feat(auth): resolver loginMember 반환

* fix(auth): token 검증 로직 interceptor로 변경

* refactor(auth): 메서드 및 에러 명 클린 코딩

* refactor(auth): 예약 테스트 리팩토링

* refactor(auth): README.md 내용 제거

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -2,12 +2,16 @@ package nextstep.reservation;
 
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
+import nextstep.support.AuthorizationException;
 import nextstep.support.DuplicateEntityException;
+import nextstep.support.NotExistEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import static nextstep.support.ErrorMessage.RESERVATION_FAIL;
+import static nextstep.support.ErrorMessage.RESERVATION_NOT_FOUND;
 
 @Service
 public class ReservationService {
@@ -49,10 +53,14 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, String name) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
-            throw new NullPointerException();
+            throw new NotExistEntityException(RESERVATION_NOT_FOUND);
+        }
+
+        if(!name.equals(reservation.getName())){
+            throw new AuthorizationException(RESERVATION_FAIL);
         }
 
         reservationDao.deleteById(id);
